[mod/#24] Order 네비게이션 menuId 타입 변경

+++ app/src/main/java/org/sopt/mcdonalds/presentation/order/navigation/OrderNavigation.kt
@@ -9,7 +9,7 @@ import kotlinx.serialization.Serializable
 import org.sopt.mcdonalds.core.common.navigation.Route
 import org.sopt.mcdonalds.presentation.order.OrderRoute
 
-fun NavController.navigateToOrder(menuId: Int, navOptions: NavOptions? = null) {
+fun NavController.navigateToOrder(menuId: Long, navOptions: NavOptions? = null) {
     navigate(Order(menuId = menuId), navOptions)
 }
 
@@ -29,5 +29,5 @@ fun NavGraphBuilder.orderGraph(
 
 @Serializable
 data class Order(
-    val menuId: Int
+    val menuId: Long
 ) : Route
